requirements.txt
RPi.GPIO
opencv-python
tflite-support
jupyter
Flask
python-logging